lab/flux-webcam/requirements.txt
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
